kerala-climate-visualizer/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
httpx==0.27.0
google-generativeai==0.8.0
python-dotenv==1.0.0
pydantic==2.7.0
